# Resolver perf: Implement TreeFS.hierarchicalLookup for getClosestPackage (2/n)