… packages/metro-file-map/src/lib/TreeFS.js